requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
chemicalx==0.1.0
class_resolver==0.4.3
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fair-esm==2.0.0
filelock==3.14.0
fonttools==4.52.4
fsspec==2024.5.0
fuzzywuzzy==0.18.0
huggingface-hub==0.23.2
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
more-itertools==10.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.44
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
pystow==0.5.4
PyTDC==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rdkit-pypi==2022.9.5
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.5.0
torch==1.11.0
torch_cluster @ file:///Users/sujaykumar/Desktop/capstone/pytorch_cluster
torch_scatter @ file:///Users/sujaykumar/Desktop/capstone/pytorch_scatter
torchdrug==0.2.1
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
